refactor: constants index 파일 삭제 및 import 경로 변경
