--- frontend/src/constants/index.ts
@@ -1,27 +1,0 @@
-export const BASE_URL = 'http://localhost:3000';
-export const BACKEND_URL = 'http://localhost:8080';
-
-export const ROUTE_PATH = {
-  MAIN: '/*',
-  EXACT_MAIN: '/',
-  SELECT_RECEIVER: '/select-receiver',
-  ENTER_COUPON: '/enter-coupon',
-  SIGN_IN: '/sign-in',
-};
-
-export const RAND_COLORS = [
-  { bg: '#FAC8BF', color: 'black' },
-  { bg: '#C0E4D9', color: 'black' },
-  { bg: 'rgb(1, 160, 230)', color: 'white' },
-  { bg: 'rgb(240, 110, 200)', color: 'white' },
-  { bg: '#FDE893', color: 'black' },
-  { bg: 'rgb(99, 51, 198)', color: 'white' },
-  { bg: '#E3C2F1', color: 'black' },
-  { bg: '#FFF970', color: 'black' },
-];
-
-export const COUPON_IMAGE = {
-  coffee:
-    'https://user-images.githubusercontent.com/41886825/177711095-cb6ff72d-d017-4c13-9a70-918f89eb6aaa.png',
-  meal: 'https://user-images.githubusercontent.com/41886825/177712393-fd41d832-bb76-44c3-a041-f619e9de9272.png',
-};
