+++ frontend/src/pages/SignIn.tsx
@@ -3,7 +3,7 @@ import GoogleIcon from '@mui/icons-material/Google';
 import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import PageLayout from '../components/@shared/PageLayout';
-import { ROUTE_PATH } from '../constants';
+import { ROUTE_PATH } from '../constants/routes';
 import { flexCenter } from '../styles/mixIn';
 import BirdLogoWhite from './../components/@shared/LogoWhite';
 import useSignIn from './../hooks/SignIn/useSignIn';
